+++ test/util/logging/logging_test.dart
@@ -170,21 +170,21 @@ void testWithLogger() {
           matches(' \\d+:\\d+:\\d+\\.\\d+\\s+💡 Info message'));
       expect(messages[3].lines.join(" "),
           matches(' \\d+:\\d+:\\d+\\.\\d+\\s+⚠️ Warning message'));
-      expect(messages[4].lines[0],
+      expect(messages[4].lines[1],
           matches('\\d+:\\d+:\\d+\\.\\d+\\s+Invalid argument\\(s\\)'));
       expect(
-          messages[4].lines[1],
+          messages[4].lines[2],
           matches(
               '\\d+:\\d+:\\d+\\.\\d+\\s+#0   Declarer.test.<anonymous closure>.<anonymous closure> \\(package:test_api/src/backend/declarer.dart:\\d+:\\d+\\)'));
-      expect(messages[4].lines[2],
+      expect(messages[4].lines[3],
           matches(' \\d+:\\d+:\\d+\\.\\d+\\s+#1   <asynchronous suspension>'));
       expect(
-          messages[4].lines[3],
+          messages[4].lines[4],
           matches(
               '\\d+:\\d+:\\d+\\.\\d+\\s+#2   StackZoneSpecification._registerUnaryCallback.<anonymous closure> \\(package:stack_trace/src/stack_zone_specification.dart\\)'));
-      expect(messages[4].lines[4],
-          matches('\\d+:\\d+:\\d+\\.\\d+\\s+#3   <asynchronous suspension>'));
       expect(messages[4].lines[5],
+          matches('\\d+:\\d+:\\d+\\.\\d+\\s+#3   <asynchronous suspension>'));
+      expect(messages[4].lines[0],
           matches(' \\d+:\\d+:\\d+\\.\\d+\\s+⛔ Error message'));
     });
     test('Test logger methods default pretty include method', () {
@@ -205,10 +205,10 @@ void testWithLogger() {
       staticLogger.d('Debug message');
       final messages = buffer.buffer.toList(growable: false);
       expect(
-          messages[0].lines[0],
+          messages[0].lines[1],
           matches(
               ' #0   OurPrettyPrinter.log \\(package:icapps_architecture/src/util/logging/impl/LoggerPrinter.dart:\\d+:\\d+\\)'));
-      expect(messages[0].lines[1], ' 🐛 Debug message');
+      expect(messages[0].lines[0], ' 🐛 Debug message');
     });
     test('Test logger methods default json', () {
       final lines = OurPrettyPrinter(
@@ -238,8 +238,8 @@ void testWithLogger() {
           .log(LogEvent(Level.error, 'Error', ArgumentError(), null));
 
       expect(
-          lines[0], '\x1B[39m\x1B[48;5;196mInvalid argument(s)\x1B[0m\x1B[49m');
-      expect(lines[1], '\x1B[38;5;196m ⛔ Error\x1B[0m');
+          lines[1], '\x1B[39m\x1B[48;5;196mInvalid argument(s)\x1B[0m\x1B[49m');
+      expect(lines[0], '\x1B[38;5;196m ⛔ Error\x1B[0m');
     });
     test('Test logger methods default error color wtf', () {
       final lines = OurPrettyPrinter(
@@ -253,8 +253,8 @@ void testWithLogger() {
           .log(LogEvent(Level.wtf, 'WTF', ArgumentError(), null));
 
       expect(
-          lines[0], '\x1B[39m\x1B[48;5;199mInvalid argument(s)\x1B[0m\x1B[49m');
-      expect(lines[1], '\x1B[38;5;199m 👾 WTF\x1B[0m');
+          lines[1], '\x1B[39m\x1B[48;5;199mInvalid argument(s)\x1B[0m\x1B[49m');
+      expect(lines[0], '\x1B[38;5;199m 👾 WTF\x1B[0m');
     });
     test('Test logger methods default build stack trace', () {
       final formatted = OurPrettyPrinter(